clean up run_flask to allow seamless running
